+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - The `Translate.translate()`-method now have overloads for improved typing.
 - User-defined ID and translation transformation framework: `id_translation.transform`
+- Bitmask translation support: `id_translation.transform.BitmaskTransformer`
 
 ### Changed
 - Make `Translator.translated_names()` optionally return a mapping dict instead of just names.